[OSS-277] Implement base lerna stuff (linking, build), create example folder & move stuff from src in it
