+++ examples/demos/src/vite-env.d.ts
