+++ package.json
@@ -8,7 +8,6 @@
   "sideEffects": false,
   "license": "MIT",
   "scripts": {
-    "dev": "vite",
     "build": "tsc && vite build",
     "cy:run": "cypress run",
     "cy:open": "cypress open",
@@ -55,10 +54,11 @@
     "eslint-plugin-react-hooks": "^4.5.0",
     "husky": "^8.0.1",
     "istanbul-lib-coverage": "^3.2.0",
+    "lerna": "^5.4.3",
     "lint-staged": "^13.0.3",
+    "nx": "^14.5.7",
     "open-cli": "^7.0.1",
     "prettier": "^2.6.2",
-    "react-router-dom": "^6.3.0",
     "rollup-plugin-copy": "^3.4.0",
     "sass": "^1.54.4",
     "typescript": "^4.6.3",
